Separate test definitions from Go tests
