+++ misc/test/azure_classic_test.go
